refactor(lyrics): split worker tests to sibling file for 1000-line cap
